config-vm: Recycle the 'process' schema
We already have two ways to specify a process to launch (for the
container process and for hooks).  This commit recycles the container
process schema for launcing the hypervisor.  I've dropped the terminal
configuration because callers are unlikely to need control over their
hypervisor's standard streams, but otherwise this is the same
structure.

The JSON Schema cheats a bit by not forbidding the terminal
properties.  We could address that if we really wanted to (JSON Schema
makes it hard to extend a previously-defined object), but I'm leaving
it to downstream tools in this commit.

Signed-off-by: W. Trevor King <wking@tremily.us>