WIP: Copy/paste new Markdown and Go for hypervisor
Michael asked for this [1], because it allows you to extend one schema
without extending the other (the terminal properties are an example of
this).  I think the overlap is much greater than the differences, so
I'd rather use the previous commit's "is is the same, except for ..."
approach.  But as far as configurations are concerned, the two
approaches are identical, so I don't really care.

I've left the JSON Schema change off this commit for now, I'll go back
through and adjust that once we have a maintainer confirming the
copy/paste approach.

[1]: http://ircbot.wl.linuxfoundation.org/meetings/opencontainers/2018/opencontainers.2018-03-07-22.00.log.html#l-53

Signed-off-by: W. Trevor King <wking@tremily.us>